SoftRend: Implement texture availability flags and fix warnings
- Added boolean flags (iIsDiffuse, iIsEmissive, etc.) to GenericFragmentShaderArgs and SoftProgram.
- Initialized GenericFragmentShaderArgs members in constructor to avoid memset warnings.
- Fixed SwGfxT::Uniform1i to correctly pass integers to SoftProgram::SetVar.
- Resolved compiler warnings in SoftProgram (GetVarSize, GetVarType, GetVar).
- Improved PBR C++ shader to utilize the new light direction and availability flags.

+++ uppsrc/api/Graphics/Cpu.cpp
@@ -121,7 +121,7 @@ template <class Gfx>
 void SwGfxT<Gfx>::Uniform1i(int idx, int i) {
 	auto& prog = Local().prog;
 	ASSERT(prog);
-	if (prog) prog->SetVar(idx, (float)i);
+	if (prog) prog->SetVar(idx, i);
 }
 
 template <class Gfx>
